More interrupt/timing fixes

+++ include/ppu/ppu.h
@@ -126,7 +126,7 @@ void initialize_ppu();
 
 void ppu_set_mirroring_mode(MirroringMode mirror_mode);
 
-uint8_t ppu_get_scanline(void);
+uint16_t ppu_get_scanline(void);
 
 uint16_t ppu_get_scanline_tick(void);
